fix compatibility issue View::class
From php-7.4.32 to php-7.4.33 there's a substancial difference. The last version makes View symfony class to return a 200 error code in the response and an error code in the json body, instead of throwing error code in the response itself. So we make such changes to maintain compatibility. Also the returned json structure is different, so we take care of any json structure received, in order to correctly handle the error.